refactor: extract shared branch picker modules
