+++ apps/web/components/task-create-dialog-repo-chip-parts.tsx
@@ -15,7 +15,7 @@ import { Tooltip, TooltipContent, TooltipTrigger } from "@kandev/ui/tooltip";
 import { Pill, type PillAction, type PillOption } from "@/components/task-create-dialog-pill";
 import type { Branch, RepositoryBranchPolicy } from "@/lib/types/http";
 import type { TaskRepoRow } from "@/components/task-create-dialog-types";
-import { computeBranchPlaceholder } from "@/components/task-create-dialog-branch-options";
+import { computeBranchPlaceholder } from "@/components/branch-picker-options";
 import {
   computeBranchDisabledReason,
   computeBranchPrefix,
