+++ apps/web/components/task-create-dialog-selectors.tsx
@@ -6,8 +6,6 @@ import { Textarea } from "@kandev/ui/textarea";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@kandev/ui/tooltip";
 import { IconPaperclip } from "@tabler/icons-react";
 import { Combobox } from "./combobox";
-import { scoreBranch } from "@/lib/utils/branch-filter";
-import { BranchRefreshButton } from "./branch-refresh-button";
 import { formatBytes } from "@/lib/utils/format-bytes";
 import {
   processFile,
@@ -48,6 +46,9 @@ import {
   useStablePluginComposerCapability,
 } from "@/lib/plugins/composer-capability";
 import { useResponsiveBreakpoint } from "@/hooks/use-responsive-breakpoint";
+
+export { BranchSelector } from "./branch-selector";
+export type { BranchOption, BranchSelectorProps } from "./branch-selector";
 
 const CURSOR_POINTER_CLASS = "cursor-pointer";
 
@@ -114,89 +115,6 @@ export const RepositorySelector = memo(function RepositorySelector({
       className={disabled ? undefined : CURSOR_POINTER_CLASS}
       triggerClassName={triggerClassName}
       testId="repository-selector"
-    />
-  );
-});
-
-type BranchOption = {
-  value: string;
-  label: string;
-  keywords?: string[];
-  renderLabel?: () => React.ReactNode;
-};
-
-type BranchSelectorProps = {
-  options: BranchOption[];
-  value: string;
-  onValueChange: (value: string) => void;
-  disabled: boolean;
-  placeholder: string;
-  searchPlaceholder: string;
-  emptyMessage: string;
-  triggerClassName?: string;
-  onRefresh?: () => void;
-  refreshing?: boolean;
-  fetchedAt?: string;
-  fetchError?: string;
-  loading?: boolean;
-  ariaLabel?: string;
-  testId?: string;
-  dropdownTestId?: string;
-  dropdownLabel?: string;
-  touchTarget?: boolean;
-  triggerId?: string;
-};
-
-export const BranchSelector = memo(function BranchSelector({
-  options,
-  value,
-  onValueChange,
-  disabled,
-  placeholder,
-  searchPlaceholder,
-  emptyMessage,
-  triggerClassName,
-  onRefresh,
-  refreshing,
-  fetchedAt,
-  fetchError,
-  loading,
-  ariaLabel,
-  testId = "branch-selector",
-  dropdownTestId,
-  dropdownLabel = t("task:baseBranch2"),
-  touchTarget = false,
-  triggerId,
-}: BranchSelectorProps) {
-  const headerAction = onRefresh ? (
-    <BranchRefreshButton
-      onRefresh={onRefresh}
-      refreshing={refreshing}
-      fetchedAt={fetchedAt}
-      fetchError={fetchError}
-      touchTarget={touchTarget}
-    />
-  ) : undefined;
-  return (
-    <Combobox
-      options={options}
-      value={value}
-      onValueChange={onValueChange}
-      placeholder={placeholder}
-      searchPlaceholder={searchPlaceholder}
-      emptyMessage={emptyMessage}
-      disabled={disabled}
-      ariaLabel={ariaLabel}
-      dropdownLabel={dropdownLabel}
-      className={disabled ? undefined : CURSOR_POINTER_CLASS}
-      triggerClassName={triggerClassName}
-      testId={testId}
-      dropdownTestId={dropdownTestId}
-      filter={scoreBranch}
-      headerAction={headerAction}
-      loading={loading}
-      touchTarget={touchTarget}
-      triggerId={triggerId}
     />
   );
 });
